Add generic runtime permission handling for Wear OS
This commit adds RequestPermissionActivity to manage runtime permission requests on Wear OS. This activity is designed to be generic, capable of requesting any Android runtime permission. It displays a rationale to the user, if needed, before prompting for system permissions and reports the outcome of the request back to the calling component.